Move driverFormLayout
